chatbot/meta/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
httpx>=0.27.0
groq>=0.9.0
gspread>=6.1.0
google-auth>=2.29.0
apscheduler>=3.10.0
python-multipart>=0.0.9

# dev
pytest>=8.2.0
pytest-asyncio>=0.23.0
